pyproject.toml
[tool.poetry]
name = "solenoid"
module = "solenoid"
version = "0.1.0"
description = "Implementation of Spring Boot Actuator and Spring Cloud Discovery, Security and Tracing"
authors = ["Grant McDonald <grantmmcdonald@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^3.13"
requests = "^2.20"
flask = "^1.0"
flask-cors = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
